+++ provider/debug/layer.go
@@ -5,7 +5,7 @@ import "github.com/terranodo/tegola/geom"
 type Layer struct {
 	name     string
 	geomType geom.Geometry
-	srid     int
+	srid     uint64
 }
 
 func (l Layer) Name() string {
@@ -16,6 +16,6 @@ func (l Layer) GeomType() geom.Geometry {
 	return l.geomType
 }
 
-func (l Layer) SRID() int {
+func (l Layer) SRID() uint64 {
 	return l.srid
 }
