+++ server/server_test.go
@@ -80,7 +80,7 @@ var testLayer3 = atlas.Layer{
 type layer struct {
 	name     string
 	geomType geom.Geometry
-	srid     int
+	srid     uint64
 }
 
 func (l layer) Name() string {
@@ -91,7 +91,7 @@ func (l layer) GeomType() geom.Geometry {
 	return l.geomType
 }
 
-func (l layer) SRID() int {
+func (l layer) SRID() uint64 {
 	return l.srid
 }
